remove JSONLD crap
