+++ phantasm-at-rest/src/test/java/com/chiralbehaviors/CoRE/phantasm/resources/WorkspaceSchemaTest.java
@@ -22,88 +22,29 @@ package com.chiralbehaviors.CoRE.phantasm.resources;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.UUID;
-import java.util.concurrent.atomic.AtomicReference;
 
 import javax.ws.rs.WebApplicationException;
 
-import org.junit.BeforeClass;
 import org.junit.Test;
 
-import com.chiralbehaviors.CoRE.domain.Product;
 import com.chiralbehaviors.CoRE.jooq.enums.ExistentialDomain;
-import com.chiralbehaviors.CoRE.kernel.phantasm.product.Argument;
-import com.chiralbehaviors.CoRE.kernel.phantasm.product.Constructor;
-import com.chiralbehaviors.CoRE.kernel.phantasm.product.InstanceMethod;
-import com.chiralbehaviors.CoRE.kernel.phantasm.product.Plugin;
-import com.chiralbehaviors.CoRE.kernel.phantasm.product.Workspace;
-import com.chiralbehaviors.CoRE.phantasm.graphql.FacetType;
-import com.chiralbehaviors.CoRE.phantasm.model.PhantasmCRUD;
-import com.chiralbehaviors.CoRE.phantasm.service.PhantasmBundle;
 import com.chiralbehaviors.CoRE.phantasm.test.location.MavenArtifact;
 import com.chiralbehaviors.CoRE.phantasm.test.product.Thing1;
 import com.chiralbehaviors.CoRE.phantasm.test.product.Thing2;
 import com.chiralbehaviors.CoRE.phantasm.test.product.Thing3;
 
 import graphql.ExecutionResult;
-import graphql.GraphQL;
-import graphql.schema.GraphQLSchema;
 
 /**
  * @author hhildebrand
  *
  */
 public class WorkspaceSchemaTest extends ThingWorkspaceTest {
-
-    private static final String COM_CHIRALBEHAVIORS_CO_RE_PHANTASM_PLUGIN_TEST = "com.chiralbehaviors.CoRE.phantasm.plugin.test";
-
-    private static ClassLoader  executionScope;
-
-    @BeforeClass
-    public static void buildExecutionScope() {
-        executionScope = PhantasmBundle.configureExecutionScope(Collections.singletonList("target/test-plugin.jar"));
-    }
-
-    @Test
-    public void testCasting() throws Exception {
-        Thing1 thing1 = model.construct(Thing1.class, ExistentialDomain.Product,
-                                        "test", "testy");
-        Thing2 thing2 = model.construct(Thing2.class, ExistentialDomain.Product,
-                                        "tester", "testier");
-        Thing3 thing3 = model.construct(Thing3.class, ExistentialDomain.Product,
-                                        "Thingy", "a favorite thing");
-        thing1.setThing2(thing2);
-        thing2.addThing3(thing3);
-
-        GraphQlResource resource = new GraphQlResource(getClass().getClassLoader());
-        Map<String, Object> variables = new HashMap<>();
-        variables.put("id", thing1.getRuleform()
-                                  .getId()
-                                  .toString());
-        variables.put("thing3", thing3.getRuleform()
-                                      .getId()
-                                      .toString());
-        QueryRequest request = new QueryRequest("mutation m($id: String!, $thing3: String!) { UpdateThing1(state: { id: $id, setThing2: $thing3}) { name } }",
-                                                variables);
-        ExecutionResult result = resource.query(null, THING_URI, request);
-        assertNotNull(result);
-
-        assertEquals(result.getErrors()
-                           .toString(),
-                     1, result.getErrors()
-                              .size());
-        assertTrue(result.getErrors()
-                         .get(0)
-                         .getMessage()
-                         .contains("ClassCastException"));
-    }
 
     @Test
     public void testGraphQlResource() throws Exception {
@@ -186,99 +127,5 @@ public class WorkspaceSchemaTest extends ThingWorkspaceTest {
                            .toString(),
                      thing3Result.get("id"));
 
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
-    public void testPlugin() throws Exception {
-
-        Workspace workspace = model.wrap(Workspace.class, scope.getWorkspace()
-                                                               .getDefiningProduct());
-        workspace.addPlugin(constructPlugin());
-
-        GraphQlResource resource = new GraphQlResource(executionScope);
-        Class<?> thing1Plugin = executionScope.loadClass(String.format("%s.Thing1_Plugin",
-                                                                       COM_CHIRALBEHAVIORS_CO_RE_PHANTASM_PLUGIN_TEST));
-        AtomicReference<String> passThrough = (AtomicReference<String>) thing1Plugin.getField("passThrough")
-                                                                                    .get(null);
-        Map<String, Object> variables = new HashMap<>();
-        variables.put("name", "hello");
-        String hello = "goodbye";
-        variables.put("description", hello);
-        QueryRequest request = new QueryRequest("mutation m ($name: String!, $description: String) { "
-                                                + "CreateThing1("
-                                                + "  state: { "
-                                                + "     setName: $name, "
-                                                + "     setDescription: $description"
-                                                + "   }) { id name description } }",
-                                                variables);
-        String bob = "Give me food or give me slack or kill me";
-        passThrough.set(bob);
-        ExecutionResult result = resource.query(null, THING_URI, request);
-
-        assertEquals(result.getErrors()
-                           .toString(),
-                     0, result.getErrors()
-                              .size());
-
-        Map<String, Object> thing1Result = (Map<String, Object>) ((Map<String, Object>) result.getData()).get("CreateThing1");
-        assertNotNull(thing1Result);
-        assertEquals(bob, thing1Result.get("description"));
-        String thing1ID = (String) thing1Result.get("id");
-        assertNotNull(thing1ID);
-        Thing1 thing1 = model.wrap(Thing1.class, model.create()
-                                                      .find(Product.class,
-                                                            UUID.fromString(thing1ID)));
-        assertEquals(bob, thing1.getDescription());
-
-        String apple = "Connie";
-        Thing2 thing2 = model.construct(Thing2.class, ExistentialDomain.Product,
-                                        apple, "Her Dobbsness");
-        thing1.setThing2(thing2);
-        variables = new HashMap<>();
-        variables.put("id", thing1ID);
-        variables.put("test", "me");
-        request = new QueryRequest("query it($id: String!, $test: String) { Thing1(id: $id) {id name instanceMethod instanceMethodWithArgument(arg1: $test) } }",
-                                   variables);
-        result = resource.query(null, THING_URI, request);
-
-        assertEquals(result.getErrors()
-                           .toString(),
-                     0, result.getErrors()
-                              .size());
-
-        thing1Result = (Map<String, Object>) ((Map<String, Object>) result.getData()).get("Thing1");
-        assertNotNull(thing1Result);
-        assertEquals(apple, thing1Result.get("instanceMethod"));
-        assertEquals("me", passThrough.get());
-        assertEquals(apple, thing1Result.get("instanceMethodWithArgument"));
-    }
-
-    private Plugin constructPlugin() throws InstantiationException {
-        Plugin testPlugin = model.construct(Plugin.class,
-                                            ExistentialDomain.Product,
-                                            "Test Plugin",
-                                            "My super green test plugin");
-        testPlugin.setFacetName("Thing1");
-        testPlugin.setPackageName(COM_CHIRALBEHAVIORS_CO_RE_PHANTASM_PLUGIN_TEST);
-        testPlugin.setConstructor(model.construct(Constructor.class,
-                                                  ExistentialDomain.Product,
-                                                  "constructor",
-                                                  "For all your construction needs"));
-        testPlugin.addInstanceMethod(model.construct(InstanceMethod.class,
-                                                     ExistentialDomain.Product,
-                                                     "instanceMethod",
-                                                     "For instance"));
-        InstanceMethod methodWithArg = model.construct(InstanceMethod.class,
-                                                       ExistentialDomain.Product,
-                                                       "instanceMethodWithArgument",
-                                                       "For all your argument needs");
-        Argument argument = model.construct(Argument.class,
-                                            ExistentialDomain.Product, "arg1",
-                                            "Who needs an argument?");
-        methodWithArg.addArgument(argument);
-        argument.setInputType("String");
-        testPlugin.addInstanceMethod(methodWithArg);
-        return testPlugin;
     }
 }
